// src/generated/repulink/types/jobState.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum JobState {
  Created,
  Funded,
  Delivered,
  Released,
  Refunded,
  Disputed,
  Resolved,
}

export type JobStateArgs = JobState;

export function getJobStateEncoder(): FixedSizeEncoder<JobStateArgs> {
  return getEnumEncoder(JobState);
}

export function getJobStateDecoder(): FixedSizeDecoder<JobState> {
  return getEnumDecoder(JobState);
}

export function getJobStateCodec(): FixedSizeCodec<JobStateArgs, JobState> {
  return combineCodec(getJobStateEncoder(), getJobStateDecoder());
}